+++ feature-odyssey/src/main/kotlin/com/hedvig/android/feature/odyssey/di/OdysseyModule.kt
@@ -1,6 +1,5 @@
 package com.hedvig.android.feature.odyssey.di
 
-import GetClaimEntryGroupUseCase
 import arrow.retrofit.adapter.either.EitherCallAdapterFactory
 import com.apollographql.apollo3.ApolloClient
 import com.hedvig.android.apollo.octopus.di.octopusClient
@@ -21,9 +20,10 @@ import com.hedvig.android.feature.odyssey.step.phonenumber.PhoneNumberViewModel
 import com.hedvig.android.feature.odyssey.step.singleitem.SingleItemViewModel
 import com.hedvig.android.feature.odyssey.step.singleitemcheckout.SingleItemCheckoutViewModel
 import com.hedvig.android.feature.odyssey.step.summary.ClaimSummaryViewModel
-import com.hedvig.android.odyssey.search.group.ClaimGroupViewModel
-import com.hedvig.android.odyssey.search.groups.ClaimGroupsViewModel
-import com.hedvig.android.odyssey.search.groups.GetNetworkClaimEntryPointGroupsUseCase
+import com.hedvig.android.feature.odyssey.search.group.ClaimGroupViewModel
+import com.hedvig.android.feature.odyssey.search.group.GetClaimEntryGroupUseCase
+import com.hedvig.android.feature.odyssey.search.groups.ClaimGroupsViewModel
+import com.hedvig.android.feature.odyssey.search.groups.GetNetworkClaimEntryPointGroupsUseCase
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import kotlinx.datetime.LocalDate
 import kotlinx.serialization.json.Json
