+++ feature-odyssey/src/main/kotlin/com/hedvig/android/feature/odyssey/search/commonclaims/SearchViewState.kt
@@ -1,6 +1,4 @@
-package com.hedvig.android.feature.odyssey.search
-
-import com.hedvig.android.odyssey.search.commonclaims.SearchableClaim
+package com.hedvig.android.feature.odyssey.search.commonclaims
 
 internal data class SearchViewState(
   val input: String? = null,
