+++ feature-odyssey/src/main/kotlin/com/hedvig/android/feature/odyssey/search/group/ClaimGroupActivity.kt
@@ -1,4 +1,4 @@
-package com.hedvig.android.odyssey.search.group
+package com.hedvig.android.feature.odyssey.search.group
 
 import android.content.Context
 import android.content.Intent
@@ -25,8 +25,8 @@ import com.hedvig.android.auth.android.AuthenticatedObserver
 import com.hedvig.android.core.designsystem.theme.HedvigTheme
 import com.hedvig.android.core.ui.appbar.CenterAlignedTopAppBar
 import com.hedvig.android.core.ui.genericinfo.GenericErrorScreen
-import com.hedvig.android.odyssey.ClaimFlowActivity
-import com.hedvig.android.odyssey.search.group.ui.ClaimGroupSelector
+import com.hedvig.android.feature.odyssey.ClaimFlowActivity
+import com.hedvig.android.feature.odyssey.search.group.ui.ClaimGroupSelector
 import org.koin.androidx.compose.koinViewModel
 import org.koin.core.parameter.parametersOf
 import slimber.log.e
